requirements.txt
streamlit
streamlit-extras
streamlit-pandas-profiling
python-dotenv
Pillow
lida
langchain-openai
langchain-experimental
pydantic
